+++ siteedit2/settings/base.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'sass_processor',
+    'core',
     'login',
     'uploader',
     'location',
